Fix : fass.cs 태도점수 데이터타입 변경
Udate fass.cs
- 데이터 테이블 태도점수 데이터타입 double -> int 타입으로 변
  확인
  fass.cs 스크립트에서 태도점수 dbManager.SaveFaceEvaluation
  로 저장시 scoreAttitude : score 캐스팅 (double) -> (int)타입으
  형변환 하도록 수정

+++ Assets/SMS/MediaPipeUnity/Samples/Scenes/Face Landmark Detection/fass.cs
@@ -383,7 +383,7 @@ public class fass : MonoBehaviour
         // currentSessionId 자동 반영 (-1 지정)
         bool success = dbManager.SaveFaceEvaluation(
             sessionId: InterviewDbManager.Instance.CurrentSessionId,
-            scoreAttitude: (double)score,
+            scoreAttitude: (int)score,
             adviceAttitudeText: adviceText,
             evalAttitudeText: summaryText
         );
